requirements.txt
anthropic
python-frontmatter
python-slugify
requests
openai
Pillow
python-dotenv
